Fix circulation import error and remove `@classmethod`
